+++ package.json
@@ -83,6 +83,7 @@
     "file-url": "^2.0.2",
     "jsdom": "^11.3.0",
     "plugin-typescript": "^7.1.0",
-    "systemjs": "^0.20.19"
+    "systemjs": "^0.20.19",
+    "unpartial": "^0.1.0"
   }
 }
