+++ src/config.ts
@@ -1,13 +1,19 @@
-import extend = require('deep-extend')
+type SystemJSConfig = SystemJSLoader.Config
 
-import { DomtureConfig } from './interfaces'
+export { SystemJSConfig }
+export interface DomtureConfig {
+  packageManager: 'npm', // | 'jspm', no support for jspm yet.
+  /**
+   * Specifies the root directory of input files.
+   */
+  rootDir: string,
+  transpiler: 'none' | 'typescript',
+  preloadScripts?: string[],
+  systemjsConfig?: SystemJSConfig
+}
 
 export const defaultConfig: DomtureConfig = {
   packageManager: 'npm',
   rootDir: '.',
   transpiler: 'none'
-}
-
-export function unpartial(config?: Partial<DomtureConfig>): DomtureConfig {
-  return extend({}, defaultConfig, config)
 }
