+++ src/interfaces.ts
@@ -3,10 +3,8 @@ import { JSDOM, DOMWindow, CookieJar, VirtualConsole } from 'jsdom'
 // re-export
 export { CookieJar, DOMWindow, VirtualConsole }
 
-type SystemJSConfig = SystemJSLoader.Config
 type SystemJS = SystemJSLoader.System
-
-export { SystemJS, SystemJSConfig }
+export { SystemJS }
 
 export interface Domture extends JSDOM {
   /**
@@ -23,15 +21,4 @@ export interface Domture extends JSDOM {
    * or relative path (`./js/pan/base/grid`)
    */
   import(identifier: string): Promise<any>
-}
-
-export interface DomtureConfig {
-  packageManager: 'npm', // | 'jspm', no support for jspm yet.
-  /**
-   * Specifies the root directory of input files.
-   */
-  rootDir: string,
-  transpiler: 'none' | 'typescript',
-  preloadScripts?: string[],
-  systemjsConfig?: SystemJSConfig
 }
